design: input 추가 & 버튼 input merge
